Initial commit.
This is the Dockerfile that was used to create a MongoDB Docker image, following the instructions located at https://docs.docker.com/engine/examples/mongodb/.

The only thing that was changed was the part of the RUN command that creates the MongoDB List file. I was receiving an error when attempting to build the image with the command as given in the Docker docs, so  changed to to use the trusty.